Refactor persistence package + implement singleton Room database
