+++ app-03-contact-manager/app/src/main/java/com/github/bitstachio/contactmanager/activity/ContactFormActivity.java
@@ -8,7 +8,7 @@ import android.widget.EditText;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.github.bitstachio.contactmanager.R;
-import com.github.bitstachio.contactmanager.db.MockDatabase;
+import com.github.bitstachio.contactmanager.persistence.MockDatabase;
 import com.github.bitstachio.contactmanager.model.Contact;
 import com.google.android.material.appbar.MaterialToolbar;
