+++ app-03-contact-manager/app/src/main/java/com/github/bitstachio/contactmanager/persistence/room/dao/IRoomContactDao.java
@@ -1,4 +1,4 @@
-package com.github.bitstachio.contactmanager.db;
+package com.github.bitstachio.contactmanager.persistence.room.dao;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
@@ -6,12 +6,13 @@ import androidx.room.Insert;
 import androidx.room.Query;
 import androidx.room.Update;
 
+import com.github.bitstachio.contactmanager.persistence.IContactDao;
 import com.github.bitstachio.contactmanager.model.Contact;
 
 import java.util.List;
 
 @Dao
-public interface RoomContactDao extends IContactDao {
+public interface IRoomContactDao extends IContactDao {
     @Query("SELECT * FROM Contact")
     List<Contact> getAll();
