+++ app-03-contact-manager/app/src/main/java/com/github/bitstachio/contactmanager/persistence/room/database/RoomDatabaseProvider.java
@@ -1,0 +1,28 @@
+package com.github.bitstachio.contactmanager.persistence.room.database;
+
+import android.content.Context;
+
+import androidx.room.Room;
+
+import lombok.Getter;
+
+@Getter
+public class RoomDatabaseProvider {
+    private static RoomDatabaseProvider instance;
+    private final ExtendedRoomDatabase database;
+
+    private RoomDatabaseProvider(Context context) {
+        database = Room.databaseBuilder(
+                context.getApplicationContext(),
+                ExtendedRoomDatabase.class,
+                "contacts.db"
+        ).build();
+    }
+
+    public static synchronized RoomDatabaseProvider getInstance(Context context) {
+        if (instance == null) {
+            instance = new RoomDatabaseProvider(context);
+        }
+        return instance;
+    }
+}
